requirements.txt
torch>=1.8.0
torchvision>=0.9.0
tqdm
Pillow
numpy
segmentation-models-pytorch
